Split refresh dispatch out of provider conformance for widget/extension targets
WidgetsExtension and Add Feed targets don't include FeedManager, so the
unified +Provider.swift extensions failed to compile there with
"Cannot find type 'FeedManager' in scope".

Move the FeedManager-dependent RefreshableFeedProvider conformance into
parallel +Refresh.swift files alongside each fetcher, and split the
registry's refreshable list into FeedProviderRegistry+Refresh.swift.

Both targets' Providers exception lists now exclude:
- RefreshableFeedProvider.swift
- FeedProviderRegistry+Refresh.swift
- {X,Instagram,YouTube}/*+Refresh.swift

The basic FeedProvider/ProfileFeedProvider conformance and the registry's
all/profileProvider/migrateAuthenticatedCookies surface remain available
in every target.
