+++ Providers/FeedProviderRegistry.swift
@@ -12,22 +12,9 @@ enum FeedProviderRegistry {
         YouTubePlaylistFetcher.self
     ]
 
-    /// Providers whose feeds use a custom refresh pipeline.
-    static let refreshable: [any RefreshableFeedProvider.Type] = [
-        XProfileFetcher.self,
-        InstagramProfileFetcher.self,
-        YouTubePlaylistFetcher.self
-    ]
-
     /// Returns the provider that owns a stored feed URL, regardless of flag state.
     static func provider(forFeedURL url: String) -> (any FeedProvider.Type)? {
         all.first { $0.matchesFeedURL(url) }
-    }
-
-    /// Returns the refreshable provider for a feed URL, or `nil` if none match
-    /// (in which case the standard RSS pipeline runs).
-    static func refreshableProvider(forFeedURL url: String) -> (any RefreshableFeedProvider.Type)? {
-        refreshable.first { $0.matchesFeedURL(url) }
     }
 
     /// Returns the profile provider matching a public profile URL.
