+++ Providers/X/XProfileFetcher+Refresh.swift
@@ -1,0 +1,19 @@
+import Foundation
+
+extension XProfileFetcher: RefreshableFeedProvider {
+
+    static func refresh(
+        feed: Feed,
+        on manager: FeedManager,
+        reloadData: Bool,
+        skipImagePreload: Bool,
+        runNLP: Bool
+    ) async throws {
+        try await manager.refreshXFeed(
+            feed,
+            reloadData: reloadData,
+            skipImagePreload: skipImagePreload,
+            runNLP: runNLP
+        )
+    }
+}
